packages/llamaindex-browseaidev/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llamaindex-browseaidev"
version = "0.1.0"
description = "LlamaIndex integration for BrowseAI Dev — verified web search with citations and confidence scores for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "BrowseAI Dev" }]
keywords = [
    "llamaindex", "llama-index", "browseaidev", "browseai-dev",
    "ai-agent", "web-search", "fact-checking", "citations",
    "confidence-scoring", "mcp", "rag", "hallucination-prevention",
    "verified-search", "evidence-based", "langchain", "crewai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "llama-index-core>=0.10.0",
    "browseaidev>=0.1.7",
]

[project.urls]
Homepage = "https://browseai.dev"
Documentation = "https://browseai.dev/docs"
Repository = "https://github.com/BrowseAI-HQ/BrowseAI-Dev"
Discord = "https://discord.gg/ubAuT4YQsT"
